added contains tab
